sw: add accuracy metric and test DRAM preloading

+++ sw/applications/src/kernels/mnist_fp64_network.h
@@ -8,16 +8,16 @@
 
 void feedforward_fp64n(uint32_t IN_CH1, uint32_t IN_CH2, uint32_t OUT_CH, 
                 double *weights, uint32_t ldW, double *biases, double* activations,
-                uint32_t ldB, double *image, uint32_t ldI, uint32_t compute_id);
+                uint32_t ldB, float *image, uint32_t ldI, uint32_t compute_id);
 
 void softmax_activation_fp64n(uint32_t IN_CH1, uint32_t IN_CH2, uint32_t OUT_CH, 
                 double *weights, uint32_t ldW, double *activations, uint32_t ldB,
-                double *image, uint32_t ldI, uint32_t compute_id, 
+                float *image, uint32_t ldI, uint32_t compute_id, 
                 uint32_t compute_num, double *max);
 
 void gradient_update_fp64(uint32_t IN_CH1, uint32_t IN_CH2, uint32_t OUT_CH, 
                 double *weight_grads, uint32_t ldW, double *bias_grads, double *activations, 
-                uint32_t ldB, double *image, uint32_t *target, uint32_t ldI, 
+                uint32_t ldB, float *image, uint32_t *target, uint32_t ldI, 
                 uint32_t compute_id, double *loss, uint32_t compute_num);
 
 void training_step_fp64(uint32_t IN_CH1, uint32_t IN_CH2, uint32_t OUT_CH,
